Included script through WebView URI
